test/requirements.txt
##### BASE REQUIREMENTS
wheel==0.29.0

# Hard dependencies:
python-magic==0.4.13
aiohttp==2.1.0
async_timeout==1.2.1

# Medium dependencies
sanic==0.5.4

# Soft dependencies (only for admin module ATM)
jinja2==2.10.1

# Probably deprecated dependencies:
requests==2.14.2

##### TEST REQUIREMENTS
coverage==4.4
flake8==3.3.0 # pyup: != 2.6.0
pytest-sugar==0.9.1
pytest-asyncio==0.6.0
requests-mock==1.3.0
#pytest-aiohttp==0.1.3


##### LOCAL DEV REQUIREMENTS
# Local development dependencies go here
Sphinx==1.5.5
autopep8==1.3.2
autoflake==0.7
isort==4.2.15
pyupio==1.0.2

# improved REPL
ipdb==0.10.3

##### Unused (only useful for PILThumb converter)
# Pillow==4.1.1
